tctl resource: convert login rule handler
